feat: model 계층 이동 application -> domain
