+++ BE/error/src/main/java/com/example/demo/schedule/domain/model/ScheduleModel.java
